NotificationSubsystem: Remove internal module, export everything
